requirements.txt
anyio==3.6.1
asgiref==3.5.2
click==8.1.3
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
numpy==1.22.4
Pillow==9.1.1
pydantic==1.9.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6